refactor: cut  all Dao about Laptop to dao directory
